remove the filter for inventories with inventory sources
the organization's name can have an /, so we need to use another character when creating the destination paths

don't export hosts that have has_inventory_sources to true

increase the max_objects limit. ToDo: make a variable for that

don't need the hosts list when the inventory has inventory_sources

don't need the hosts list when the inventory has inventory_sources

the organization's name can have an /, so we need to use another character when creating the destination paths

DO need the hosts list when the inventory has inventory_sources

DO NOT need the hosts list when the inventory has inventory_sources. Unaffordable by now: out of resources at large environments